refactor(cli): extract shared range chunking, add integer and offset planners
First slice of #479. `faucet backfill` already knew how to split a time range
into contiguous, DST-correct windows; the `partition:` block needs the same
thing plus integer and offset ranges. Two independent chunkers that disagreed on
boundary handling is exactly the bug this feature must not ship, so the window
logic moves into `cli/src/chunking.rs` and `backfill::plan` re-exports it.

The re-export is the point: every existing backfill call site and test keeps
compiling and passing against the moved code, which is what demonstrates the move
did not change planning behaviour. Backfill's 27 unit tests and 2 end-to-end
tests pass unchanged, and the window tests moved with the code they cover.

This lives in the CLI rather than `faucet-core`, contrary to the issue's original
sketch. Both consumers are CLI-layer — the substitution that uses these chunks
walks config strings, which no connector participates in — and putting it in core
would mean making `chrono` non-optional there and adding `chrono-tz`, neither of
which a connector author needs.

New planners:

  - `plan_int_chunks(from, to, chunk_size, bounds)` — splits an integer range.
  - `plan_offset_chunks(total, chunk_size)` — splits a countable result set into
    offset/limit slices. Takes a **count**, never a maximum key: chunking an id
    range from a count is silently wrong the moment ids are sparse.

`Bounds` has no default on purpose. With `chunk_size: 10000` from 0, inclusive
gives `[0, 9999]` and half-open gives `[0, 10000)`. Half-open chunks against a
source whose upper bound is inclusive fetch record 10000 **twice**; inclusive
chunks against an exclusive source **never fetch** record 9999. Neither surfaces
as an error, so the user has to say which their source is.

The tests pin that as a property rather than a few examples: both bounds tile
their range exactly once — no gap, no overlap — across 72 combinations of start,
span, and chunk size, plus i64 boundary ranges that would wrap in i64 arithmetic
(the span is computed in i128).

Addresses #479
